+++ .changeset/shy-avocados-sip.md
@@ -1,5 +1,4 @@
 ---
-'@thorswap-lib/swapkit-explorers': patch
 '@thorswap-lib/toolbox-evm': patch
 '@thorswap-lib/walletconnect': patch
 '@thorswap-lib/swapkit-core': patch
