+++ packages/swapkit/swapkit-core/package.json
@@ -5,7 +5,6 @@
     "@ethersproject/units": "5.7.0",
     "@thorswap-lib/helpers": "workspace:*",
     "@thorswap-lib/swapkit-entities": "workspace:*",
-    "@thorswap-lib/swapkit-explorers": "workspace:*",
     "@thorswap-lib/types": "workspace:*",
     "bignumber.js": "9.1.1"
   },
@@ -53,8 +52,8 @@
     "build": "echo 'Build swapkit-core 🛠';NODE_OPTIONS=--max_old_space_size=16384 pnpm vite build && echo 'Build succeeded swapkit-core ✅'",
     "clean": "rm -rf dist .turbo",
     "lint": "eslint ./ --ext .ts,.tsx --fix; tsc --noEmit",
-    "test": "echo 'vitest --run'",
-    "test:coverage": "echo 'vitest run --coverage'"
+    "test": "vitest --run",
+    "test:coverage": "vitest run --coverage"
   },
   "type": "module",
   "types": "./dist/index.d.ts",
