--- packages/swapkit/swapkit-explorers/vite.config.ts
@@ -1,14 +1,0 @@
-import thorswapViteConfig from '@internal/config';
-import { resolve } from 'path';
-
-import { name } from './package.json';
-
-const viteConfig = thorswapViteConfig(name, {
-  build: {
-    lib: {
-      entry: resolve(__dirname, 'src/index.ts'),
-    },
-  },
-});
-
-export default viteConfig;
